+++ lib/models/money_objects/accounts/account.dart
@@ -13,6 +13,8 @@ export 'package:money/models/money_objects/accounts/account_types.dart';
 class Account extends MoneyObject {
   @override
   int get uniqueId => id.value;
+  @override
+  set uniqueId(value) => id.value = value;
 
   // Id
   // 0|Id|INT|0||1