+++ lib/models/money_objects/currencies/currency.dart
@@ -17,6 +17,8 @@ import 'package:money/models/money_objects/money_objects.dart';
 class Currency extends MoneyObject {
   @override
   int get uniqueId => id.value;
+  @override
+  set uniqueId(value) => id.value = value;
 
   // 0
   FieldId<Currency> id = FieldId<Currency>(